Create utility and move formatLocalTime function
